requirements.txt
pyserial
pylint


